break up utils.misc to fix circular imports
