+++ packages/utils/src/index.ts
@@ -1,14 +1,17 @@
 export * from './async';
+export * from './browser';
+export * from './dsn';
 export * from './error';
 export * from './is';
 export * from './logger';
 export * from './memo';
 export * from './misc';
+export * from './node';
 export * from './object';
 export * from './path';
 export * from './promisebuffer';
+export * from './stacktrace';
 export * from './string';
 export * from './supports';
 export * from './syncpromise';
 export * from './instrument';
-export * from './dsn';
