+++ packages/utils/src/object.ts
@@ -1,9 +1,10 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { ExtendedError, WrappedFunction } from '@sentry/types';
 
+import { htmlTreeAsString } from './browser';
 import { isElement, isError, isEvent, isInstanceOf, isPlainObject, isPrimitive, isSyntheticEvent } from './is';
 import { Memo } from './memo';
-import { getFunctionName, htmlTreeAsString } from './misc';
+import { getFunctionName } from './stacktrace';
 import { truncate } from './string';
 
 /**
